Add events command to container
