+++ src/commands/events.rs
@@ -1,11 +1,9 @@
-use crate::utils;
-use cgroups::common;
 use clap::Clap;
-use std::{path::PathBuf, thread, time::Duration};
+use std::path::PathBuf;
 
 use anyhow::{bail, Context, Result};
 
-use crate::container::{Container, ContainerStatus};
+use crate::container::Container;
 
 #[derive(Clap, Debug)]
 pub struct Events {
@@ -28,36 +26,9 @@ impl Events {
             bail!("{} doesn't exist.", self.container_id)
         }
 
-        let container = Container::load(container_dir)?.refresh_status()?;
-        if !container.state.status.eq(&ContainerStatus::Running) {
-            bail!("{} is not in running state", self.container_id);
-        }
-
-        let cgroups_path = utils::get_cgroup_path(
-            &container
-                .spec()?
-                .linux
-                .context("no linux in spec")?
-                .cgroups_path,
-            &self.container_id,
-        );
-        let use_systemd = container
-            .systemd()
-            .context("Could not determine cgroup manager")?;
-
-        let cgroup_manager = common::create_cgroup_manager(cgroups_path, use_systemd)?;
-        match self.stats {
-            true => {
-                let stats = cgroup_manager.stats()?;
-                println!("{}", serde_json::to_string_pretty(&stats)?);
-            }
-            false => loop {
-                let stats = cgroup_manager.stats()?;
-                println!("{}", serde_json::to_string_pretty(&stats)?);
-                thread::sleep(Duration::from_secs(self.interval as u64));
-            },
-        }
-
-        Ok(())
+        let mut container = Container::load(container_dir)?;
+        container
+            .events(self.interval, self.stats)
+            .with_context(|| format!("failed to get events from container {}", self.container_id))
     }
 }
